sessiondata: move LocalOnlySessionData enums to sessiondatapb
This is in preparation for enum-ifying these variables.
All changes applied here are through Goland refactor.

Release note: None
