+++ pkg/sql/logictest/BUILD.bazel
@@ -29,7 +29,7 @@ go_library(
         "//pkg/sql/randgen",
         "//pkg/sql/row",
         "//pkg/sql/sem/tree",
-        "//pkg/sql/sessiondata",
+        "//pkg/sql/sessiondatapb",
         "//pkg/sql/stats",
         "//pkg/testutils",
         "//pkg/testutils/floatcmp",
@@ -71,7 +71,7 @@ go_test(
         "//pkg/security/securitytest",
         "//pkg/server",
         "//pkg/sql",
-        "//pkg/sql/sessiondata",
+        "//pkg/sql/sessiondatapb",
         "//pkg/sql/stats",
         "//pkg/sql/tests",
         "//pkg/testutils/serverutils",
