+++ pkg/sql/schema_changer_state.go
@@ -12,12 +12,12 @@ package sql
 
 import (
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
+	"github.com/cockroachdb/cockroach/pkg/sql/sessiondatapb"
 )
 
 // SchemaChangerState is state associated with the new schema changer.
 type SchemaChangerState struct {
-	mode  sessiondata.NewSchemaChangerMode
+	mode  sessiondatapb.NewSchemaChangerMode
 	state scpb.State
 	// stmts contains the SQL statements involved in the schema change. This is
 	// the bare minimum of statement information we need for testing, but in the
