+++ pkg/sql/sessiondatapb/BUILD.bazel
@@ -4,7 +4,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "sessiondatapb",
-    srcs = ["session_data.go"],
+    srcs = [
+        "local_only_session_data.go",
+        "session_data.go",
+    ],
     embed = [":sessiondatapb_go_proto"],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/sessiondatapb",
     visibility = ["//visibility:public"],
