+++ pkg/sql/vars.go
@@ -437,7 +437,7 @@ var varGen = map[string]sessionVar{
 	// CockroachDB extension.
 	`distsql`: {
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			mode, ok := sessiondata.DistSQLExecModeFromString(s)
+			mode, ok := sessiondatapb.DistSQLExecModeFromString(s)
 			if !ok {
 				return newVarValueError(`distsql`, s, "on", "off", "auto", "always", "2.0-auto", "2.0-off")
 			}
@@ -448,7 +448,7 @@ var varGen = map[string]sessionVar{
 			return evalCtx.SessionData.DistSQLMode.String()
 		},
 		GlobalDefault: func(sv *settings.Values) string {
-			return sessiondata.DistSQLExecMode(DistSQLClusterExecMode.Get(sv)).String()
+			return sessiondatapb.DistSQLExecMode(DistSQLClusterExecMode.Get(sv)).String()
 		},
 	},
 
@@ -477,7 +477,7 @@ var varGen = map[string]sessionVar{
 	`experimental_distsql_planning`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`experimental_distsql_planning`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			mode, ok := sessiondata.ExperimentalDistSQLPlanningModeFromString(s)
+			mode, ok := sessiondatapb.ExperimentalDistSQLPlanningModeFromString(s)
 			if !ok {
 				return newVarValueError(`experimental_distsql_planning`, s,
 					"off", "on", "always")
@@ -489,7 +489,7 @@ var varGen = map[string]sessionVar{
 			return evalCtx.SessionData.ExperimentalDistSQLPlanningMode.String()
 		},
 		GlobalDefault: func(sv *settings.Values) string {
-			return sessiondata.ExperimentalDistSQLPlanningMode(experimentalDistSQLPlanningClusterMode.Get(sv)).String()
+			return sessiondatapb.ExperimentalDistSQLPlanningMode(experimentalDistSQLPlanningClusterMode.Get(sv)).String()
 		},
 	},
 
@@ -750,7 +750,7 @@ var varGen = map[string]sessionVar{
 	// CockroachDB extension.
 	`serial_normalization`: {
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			mode, ok := sessiondata.SerialNormalizationModeFromString(s)
+			mode, ok := sessiondatapb.SerialNormalizationModeFromString(s)
 			if !ok {
 				return newVarValueError(`serial_normalization`, s,
 					"rowid", "virtual_sequence", "sql_sequence", "sql_sequence_cached")
@@ -762,7 +762,7 @@ var varGen = map[string]sessionVar{
 			return evalCtx.SessionData.SerialNormalizationMode.String()
 		},
 		GlobalDefault: func(sv *settings.Values) string {
-			return sessiondata.SerialNormalizationMode(
+			return sessiondatapb.SerialNormalizationMode(
 				SerialNormalizationMode.Get(sv)).String()
 		},
 	},
@@ -1400,7 +1400,7 @@ var varGen = map[string]sessionVar{
 	`experimental_use_new_schema_changer`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`experimental_use_new_schema_changer`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			mode, ok := sessiondata.NewSchemaChangerModeFromString(s)
+			mode, ok := sessiondatapb.NewSchemaChangerModeFromString(s)
 			if !ok {
 				return newVarValueError(`experimental_use_new_schema_changer`, s,
 					"off", "on", "unsafe_always")
@@ -1412,7 +1412,7 @@ var varGen = map[string]sessionVar{
 			return evalCtx.SessionData.NewSchemaChangerMode.String()
 		},
 		GlobalDefault: func(sv *settings.Values) string {
-			return sessiondata.NewSchemaChangerMode(experimentalUseNewSchemaChanger.Get(sv)).String()
+			return sessiondatapb.NewSchemaChangerMode(experimentalUseNewSchemaChanger.Get(sv)).String()
 		},
 	},
